requirements_py38.txt
ansicolors==1.1.8
docker==5.0.2
h5py==2.10.0
matplotlib==3.3.4
numpy==1.19.5
pyyaml==5.4
psutil==5.6.6
scipy==1.5.4
scikit-learn
jinja2==2.10.1
pandas==1.1.5
